chore: make all types that are used or returned by the public interface also public
Based on warnings produced by typedoc.
