+++ src/featureMatching/descriptors/getBriefDescriptors.ts
@@ -3,12 +3,10 @@ import { GaussianBlurSigmaOptions } from '../../filters';
 import checkProcessable from '../../utils/validators/checkProcessable';
 import { OrientedFastKeypoint } from '../keypoints/getOrientedFastKeypoints';
 import { compareIntensity } from '../utils/compareIntensity';
-import {
-  getGaussianPoints,
-  GetGaussianPointsOptions,
-} from '../utils/getGaussianPoints';
+import { getGaussianPoints } from '../utils/getGaussianPoints';
 
 import { getKeypointPatch } from './utils/getKeypointPatch';
+import { GetGaussianPointsOptions } from '../../utils/utils.types';
 
 export interface GetBriefDescriptorsOptions {
   /**
