+++ src/featureMatching/visualize/drawMatches.ts
@@ -1,12 +1,13 @@
 import { Image } from '../../Image';
 import { FastKeypoint } from '../keypoints/getFastKeypoints';
 import { Match } from '../matching/bruteForceMatch';
-import { getColors, GetColorsOptions } from '../utils/getColors';
+import { getColors } from '../utils/getColors';
 import { getMatchColor } from '../utils/getMatchColor';
 import { sortByDistance } from '../utils/sortByDistance';
 
 import { Montage } from './Montage';
 import { scaleKeypoints } from './scaleKeypoints';
+import { GetColorsOptions } from '../featureMatching.types';
 
 export interface DrawMatchesOptions {
   /**
