+++ src/index.ts
@@ -9,10 +9,12 @@ export * from './geometry';
 export * from './Image';
 export * from './load';
 export * from './Mask';
+export * from './maskAnalysis';
 export * from './morphology';
 export * from './operations';
 export * from './roi';
 export * from './save';
+export * from './utils/utils.types';
 export * from './utils/constants/channelLabels';
 export * from './utils/constants/colorModels';
 export { BorderType } from './utils/interpolateBorder';
