+++ src/load/index.ts
@@ -2,6 +2,7 @@ export * from './decode';
 export * from './decodeJpeg';
 export * from './decodePng';
 export * from './decodeTiff';
+export * from './load.types';
 export * from './read';
 export * from './readCanvas';
 export * from './readImg';
