+++ src/load/load.types.ts
@@ -1,0 +1,7 @@
+export interface ImageMetadata {
+  tiff: {
+    fields: Map<number, unknown>;
+    tags: Record<string, unknown>;
+  };
+  exif: Record<string, unknown>;
+}
