+++ src/maskAnalysis/index.ts
@@ -1,4 +1,1 @@
-export * from './getFeret';
-export * from './getBorderPoints';
-export * from './getConvexHull';
-export * from './getMbr';
+export * from './maskAnalysis.types';
