+++ src/operations/index.ts
@@ -7,5 +7,6 @@ export * from './threshold';
 export * from './grey';
 export * from './copyTo';
 export * from './crop';
+export * from './operations.types';
 export * from './paintMaskOnImage';
 export * from './paintMaskOnMask';
