+++ src/operations/operations.types.ts
@@ -1,0 +1,6 @@
+export interface CropAlphaOptions {
+  /**
+   * Threshold from which rows and columns should be kept.
+   */
+  threshold?: number;
+}
