+++ src/roi/Roi.ts
@@ -1,32 +1,22 @@
 import { Mask } from '../Mask';
-import {
-  GetBorderPointsOptions,
-  getFeret,
-  Feret,
-  getConvexHull,
-  getMbr,
-  Mbr,
-} from '../maskAnalysis';
+
 import { Point } from '../utils/geometry/points';
 
 import { RoiMap } from './RoiMapManager';
 import { getBorderPoints } from './getBorderPoints';
 import { getMask, GetMaskOptions } from './getMask';
-import { Ellipse, getEllipse } from './properties/getEllipse';
+import { getEllipse } from './properties/getEllipse';
+import { Feret, GetBorderPointsOptions, Mbr } from '../maskAnalysis';
+import { Border, Ellipse } from './roi.types';
+import { getConvexHull } from '../maskAnalysis/getConvexHull';
+import { getMbr } from '../maskAnalysis/getMbr';
+import { getFeret } from '../maskAnalysis/getFeret';
+
 /**
  * Properties of borders of ROI.
  *
  */
-interface Border {
-  /**
-   * Refers  to the roiID of the contiguous ROI.
-   */
-  connectedID: number;
-  /**
-   * Length of the border with connectedID.
-   */
-  length: number;
-}
+
 interface Computed {
   perimeter: number;
   borders: Border[]; // external and internal ids which are not equal to the current roi ID
@@ -393,8 +383,7 @@ export class Roi {
    */
   get ellipse(): Ellipse {
     return this.#getComputed('ellipse', () => {
-      const ellipse = getEllipse(this);
-      return ellipse;
+      return getEllipse(this);
     });
   }
